stock_ai_agent/requirements.txt
python-telegram-bot==20.7
pandas
numpy
yfinance
matplotlib
